backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
numpy==1.26.0
pillow==10.4.0
tqdm==4.67.1
torch==2.6.0
torchvision==0.21.0
gunicorn==23.0.0
matplotlib==3.8.0
bcrypt==4.2.1
pyjwt[crypto]
python-dotenv
cryptography